composite build instead of buildsrc and fix review remarks
