+++ build-logic/src/main/kotlin/org.somda.sdc/XjcPlugin.kt
