+++ build.gradle.kts
@@ -1,6 +1,3 @@
-val defaultVersion = "9.1.0-SNAPSHOT"
-val actualVersion = project.findProperty("revision") ?: defaultVersion
-val actualRevision = project.findProperty("changelist") ?: ""
-
-group = "com.draeger.medical"
-version = "$actualVersion$actualRevision"
+plugins {
+    id("com.draeger.medical.java-conventions")
+}
