+++ settings.gradle.kts
@@ -2,7 +2,15 @@
  * This file was generated by the Gradle 'init' task.
  */
 
+pluginManagement {
+    // Include 'plugins build' to define convention plugins.
+    includeBuild("build-logic")
+}
+
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+
 rootProject.name = "sdccc_parent"
 include(":biceps-model")
 include(":dpws-model")
 include(":sdccc")
+
